feat:完成传输接口的getTransferLogs获取传输历史记录方法、getTransferSummary获取传输统计摘要方法、更新下载传输状态updateTransferStatus接口

+++ src/main/java/com/liteisle/common/domain/TransferLog.java
@@ -67,7 +67,7 @@ public class TransferLog {
     /**
      * 传输总耗时，单位：毫秒 (ms)
      */
-    private Integer transferDurationMs;
+    private Long transferDurationMs;
 
     /**
      * 发起传输的客户端IP地址，用于安全审计
